--- backend/dguonoff/src/main/java/backend/spectrum/dguonoff/domain/user/controller/UserController.java
@@ -1,5 +1,0 @@
-package backend.spectrum.dguonoff.domain.user.controller;
-
-public class UserController {
-
-}
