--- backend/dguonoff/src/main/java/backend/spectrum/dguonoff/domain/user/domain/Role.java
@@ -1,5 +1,0 @@
-package backend.spectrum.dguonoff.domain.user.domain;
-
-public enum Role {
-    NORMAL, ADMIN, MASTER
-}
