+++ backend/dguonoff/src/main/java/backend/spectrum/dguonoff/domain/user/entity/Role.java
@@ -1,0 +1,5 @@
+package backend.spectrum.dguonoff.domain.user.entity;
+
+public enum Role {
+    NORMAL, ADMIN, MASTER
+}
