+++ backend/dguonoff/src/main/java/backend/spectrum/dguonoff/domain/user/repository/UserRepository.java
@@ -1,6 +1,6 @@
 package backend.spectrum.dguonoff.domain.user.repository;
 
-import backend.spectrum.dguonoff.domain.user.domain.User;
+import backend.spectrum.dguonoff.domain.user.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Modifying;
 import org.springframework.data.jpa.repository.Query;
